+++ cx-portal/src/components/overlays/AddUser/AddUserDeny.tsx
@@ -26,7 +26,7 @@ import {
   DialogHeader,
 } from 'cx-portal-shared-components'
 import { useDispatch } from 'react-redux'
-import { closeOverlay } from 'features/control/overlay/actions'
+import { closeOverlay } from 'features/control/overlay'
 import { useNavigate } from 'react-router-dom'
 import { PAGES } from 'types/Constants'
 import { IdentityProvider } from 'features/admin/idpApiSlice'