+++ cx-portal/src/components/overlays/AppOverViewConfirm/AppDetailsOverlay.tsx
@@ -25,7 +25,7 @@ import {
 } from 'cx-portal-shared-components'
 import { useTranslation } from 'react-i18next'
 import AppOverViewDetails from 'components/pages/AppOverview/AppOverViewDetails'
-import { closeOverlay } from 'features/control/overlay/actions'
+import { closeOverlay } from 'features/control/overlay'
 import { useFetchAppStatusQuery } from 'features/appManagement/apiSlice'
 import { useDispatch } from 'react-redux'