+++ cx-portal/src/features/admin/index.ts
@@ -18,23 +18,13 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { OVERLAYS } from 'types/Constants'
+import { combineReducers } from 'redux'
+import { slice as registration } from './registration/slice'
+import { slice as user } from './userDeprecated/slice'
+import { slice as userOwn } from './userOwn/slice'
 
-export const name = 'control/overlay'
-
-export type OverlayState = {
-  type: OVERLAYS
-  id: string
-  title?: string
-  status?: boolean
-  subTitle?: string
-  roles?: string[]
-}
-
-export const initialState = {
-  type: OVERLAYS.NONE,
-  id: '',
-  title: '',
-  displayName: '',
-  subTitle: '',
-}
+export default combineReducers({
+  user: user.reducer,
+  registration: registration.reducer,
+  userOwn: userOwn.reducer,
+})